refactor: event handlers in `background.ts` (#49)
